requirements.txt
beautifulsoup4~=4.12.3
httpx~=0.28.1
httpx[http2]~=0.28.1
PyYaml~=6.0
